prefer uid_t/gid_t instead of int/uint where possible
svn path=/trunk/kdeadmin/kuser/; revision=325488

+++ propdlg.cpp
@@ -834,7 +834,7 @@ void propdlg::slotOk()
   
   bool one = ( mUsers.getFirst() == mUsers.getLast() );
 
-  uid_t newuid = leid->text().toInt();
+  uid_t newuid = leid->text().toULong();
 
   if ( one && olduid != newuid )
   {
